Update //
The SDK itself is a free download, but beta SDKs require a paid
developer account. In order to release software, one must enroll in the
iPhone Developer Program—a step requiring payment and Apple's approval.
As of February 2015, cost of enrollment in the iPhone Developer Program
is US$99 per year (the cost varies from country to country) for the
standard program. Signed keys are provided to upload the application to
Apple's App Store. Applications can be distributed in three ways:
through the App Store, through enterprise deployment to a company's
employees only, and on an "Ad-hoc" basis to up to 100 iPhones. Once
distributed through the App Store, a developer can request up to 50
promotional codes that can be used to freely distribute a commercial
application he or she has developed. The SDK can only be installed in
OS X, and the latest SDK (iOS SDK 8.0 with Xcode 6.0.1, as of September
2014) requires OS X 10.9.4 "Mavericks" or later.